requirements.txt
azure-functions
azure-storage-blob
requests
PyGithub
